refactor: BillStep hook 분리
