+++ client/src/pages/BillPage/steps/MembersStep.tsx
@@ -1,20 +1,13 @@
 import {css} from '@emotion/react';
-import {useEffect, useState} from 'react';
-import {useNavigate} from 'react-router-dom';
-
-import useRequestPostMembers from '@hooks/queries/member/useRequestPostMembers';
-import useRequestPostBill from '@hooks/queries/bill/useRequestPostBill';
 import Top from '@components/Design/components/Top/Top';
 import ChipButton from '@components/Design/components/ChipButton/ChipButton';
 import {Member} from 'types/serviceType';
 
 import {FixedButton, Flex, LabelInput, Text} from '@components/Design';
 
-import getEventIdByUrl from '@utils/getEventIdByUrl';
-
-import REGEXP from '@constants/regExp';
-
-import {BillInfo, BillStep} from '../AddBillFunnel';
+import {BillInfo} from '../AddBillFunnel';
+import useMembersStep from '@hooks/useMembersStep';
+import {BillStep} from '@hooks/useAddBillFunnel';
 
 interface Props {
   billInfo: BillInfo;
@@ -23,98 +16,18 @@ interface Props {
   currentMembers: Member[];
 }
 
-const MembersStep = ({billInfo, setBillInfo, setStep, currentMembers}: Props) => {
-  const [errorMessage, setErrorMessage] = useState('');
-  const [nameInput, setNameInput] = useState('');
-  const [canRequestPostBill, setCanRequestPostBill] = useState(false);
-
+const MembersStep = ({billInfo, setBillInfo, currentMembers, setStep}: Props) => {
   const {
-    postMembers,
-    responseMemberIds,
-    isSuccess: isSuccessPostMembers,
-    isPending: isPendingPostMembers,
-  } = useRequestPostMembers();
-  const {postBill, isSuccess: isSuccessPostBill, isPending: isPendingPostBill} = useRequestPostBill();
-  const navigate = useNavigate();
-  const eventId = getEventIdByUrl();
-
-  const onNameInputChange = (value: string) => {
-    if (REGEXP.memberName.test(value)) {
-      setNameInput(value);
-    }
-  };
-
-  const handleNameInputChange = (event: React.ChangeEvent<HTMLInputElement>) => {
-    if (event.target.value.length > 4) {
-      setErrorMessage('이름은 4자까지 입력 가능해요');
-      onNameInputChange(nameInput.slice(0, 4));
-    } else {
-      setErrorMessage('');
-      onNameInputChange(event.target.value);
-    }
-  };
-
-  const canAddMembers = nameInput && !errorMessage;
-
-  const canSubmitMembers = billInfo.members.length !== 0;
-
-  const setBillInfoMemberWithId = (name: string) => {
-    const existingMember = currentMembers.find(currentMember => currentMember.name === name);
-    if (existingMember) {
-      setBillInfo(prev => ({...prev, members: [...prev.members, {id: existingMember.id, name: name}]}));
-    } else {
-      setBillInfo(prev => ({...prev, members: [...prev.members, {id: -1, name: name}]}));
-    }
-  };
-
-  const handleNameInputEnter = (event: React.KeyboardEvent<HTMLInputElement>) => {
-    if (event.nativeEvent.isComposing) {
-      return;
-    }
-    if (event.key === 'Enter' && canAddMembers) {
-      event.preventDefault();
-      if (!billInfo.members.map(({name}) => name).includes(nameInput)) {
-        setBillInfoMemberWithId(nameInput);
-      }
-      setNameInput('');
-    }
-  };
-
-  const setStepTitle = () => {
-    setStep('title');
-  };
-
-  const handlePostBill = async () => {
-    if (billInfo.members.map(({id}) => id).includes(-1)) {
-      postMembers({
-        members: billInfo.members.filter(member => member.id === -1),
-      });
-    } else {
-      postBill({
-        title: billInfo.title,
-        price: Number(billInfo.price.replace(',', '')),
-        members: billInfo.members.map(({id}) => id),
-      });
-    }
-  };
-
-  useEffect(() => {
-    if (isSuccessPostMembers && responseMemberIds) {
-      postBill({
-        title: billInfo.title,
-        price: Number(billInfo.price.replace(',', '')),
-        members: billInfo.members.map(member =>
-          member.id === -1 ? responseMemberIds.members.find(m => m.name === member.name)?.id || member.id : member.id,
-        ),
-      });
-    }
-  }, [isSuccessPostMembers, responseMemberIds]);
-
-  useEffect(() => {
-    if (isSuccessPostBill) {
-      navigate(`/event/${eventId}/admin`);
-    }
-  }, [isSuccessPostBill]);
+    errorMessage,
+    nameInput,
+    handleNameInputChange,
+    handleNameInputEnter,
+    isPendingPostBill,
+    isPendingPostMembers,
+    canSubmitMembers,
+    handlePostBill,
+    handlePrevStep,
+  } = useMembersStep({billInfo, setBillInfo, currentMembers, setStep});
 
   return (
     <>
@@ -176,7 +89,7 @@ const MembersStep = ({billInfo, setBillInfo, setStep, currentMembers}: Props) =>
         variants={isPendingPostBill || isPendingPostMembers ? 'loading' : 'primary'}
         disabled={!canSubmitMembers}
         onClick={handlePostBill}
-        onBackClick={setStepTitle}
+        onBackClick={handlePrevStep}
       >
         추가완료
       </FixedButton>
